+++ index.html
@@ -232,7 +232,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src = "function.js"></script>
-  <script src = "main.js"></script>
+  <script type = 'module' src = "./js/main.js"></script>
 </body>
 </html>
